requirements.txt
DateTime==4.0
Django==1.4.2
MySQL-python==1.2.4
South==0.8.1
Unidecode==0.04.12
amqp==1.0.11
anyjson==0.3.3
billiard==2.7.3.28
celery==3.0.19
distribute==0.6.43
django-celery==3.0.17
django-kombu==0.9.4
facebook-sdk==0.4.0
facepy==0.8.4
kombu==2.5.10
mechanize==0.2.5
python-dateutil==2.1
pytz==2013b
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
zope.interface==4.0.5
